use a service for mock-login
