+++ packages/mock-auth/addon/templates/components/mock-login.hbs
@@ -1,1 +1,1 @@
-{{yield (perform login) }}
+{{yield (perform mockLogin.login) }}
