+++ packages/mock-auth/app/services/mock-login.js
@@ -1,0 +1,1 @@
+export { default } from '@cardstack/mock-auth/services/mock-login';
